feat: 過多な設計を見直し
